refactor: 확장성 고려 및 역할 분리를 위한 재구조화
